+++ src/index.ts
@@ -18,7 +18,6 @@ const demoMode =                parseParam(process.env.DEMO_MODE, false);
 const enableDebugger =          parseParam(process.env.ENABLE_DEBUGGER, true);
 const maxMemory =               parseParam(process.env.MAX_MEMOMORY_PERCENT, 99);
 const maxCPU =                  parseParam(process.env.MAX_CPU_PERCENT, 99);
-const autoQueue =               parseParam(process.env.AUTO_SCALE, false);
 const keepAlive =               parseParam(process.env.KEEP_ALIVE, false);
 const chromeRefreshTime =       parseParam(process.env.CHROME_REFRESH_TIME, thirtyMinutes);
 const maxChromeRefreshRetries = parseParam(process.env.MAX_CHROME_REFRESH_RETRIES, 5);
@@ -30,7 +29,6 @@ const healthFailureURL =        process.env.FAILED_HEALTH_URL || null;
 const metricsJSONPath =         process.env.METRICS_JSON_PATH || null;
 
 new BrowserlessServer({
-  autoQueue,
   chromeRefreshTime,
   connectionTimeout,
   demoMode,
